pyproject.toml
[tool.poetry]
name = "smartreportgenerator"
version = "0.1.0"
description = ""
authors = ["Emre Tuncer <emretuncer.developer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
pillow = "^11.1.0"
fpdf2 = "^2.8.2"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
